Run password removal tests (#6300)
